moved save commit Id method to IdeContext
