+++ cli/src/main/java/com/devonfw/tools/ide/commandlet/AbstractUpdateCommandlet.java
@@ -133,7 +133,7 @@ public abstract class AbstractUpdateCommandlet extends Commandlet {
         }
         gitContext.pullOrClone(GitUrl.of(repository), settingsPath);
       }
-      gitContext.saveCurrentCommitId(settingsPath, this.context.getSettingsCommitIdPath());
+      this.context.saveCurrentCommitId(settingsPath, this.context.getSettingsCommitIdPath());
       step.success("Successfully updated settings repository.");
     } finally {
       if (step != null) {
